refactor(tests): extract shared library-query and eval assertion helpers
Reduce cross-file/within-file duplication in tests and evals:
- library-fixture.ts: add STALENESS_RISK fixture + expectQueryDegradesOnBrokenDb
  helper, consumed by find-duplicates/find-similar query tests
- eval scenario defs: extract rewriteAssertionHead()/createClipAssertionHead()
  for the repeated assertion heads

Evals duplication 0.44% -> 0%, tests 0.95% -> 0.93%. Remaining cross-file test
clones are vi.mock hoisting boilerplate (factories can't reference imports).
