+++ src/mcp-server/live-library/tests/query/find-similar.test.ts
@@ -6,8 +6,9 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { findSimilar } from "../../query/find-similar.ts";
 import {
-  createBrokenLibraryDb,
+  expectQueryDegradesOnBrokenDb,
   setupLibraryFixtureLifecycle,
+  STALENESS_RISK,
 } from "../fixtures/library-fixture.ts";
 
 vi.mock(import("../../live-db-path.ts"), () => ({
@@ -99,19 +100,13 @@ describe("findSimilar", () => {
   });
 
   it("surfaces a staleness advisory from the DB layer", async () => {
-    const risk = {
-      kind: "wal-pending" as const,
-      dbMtime: 1,
-      walMtime: 2,
-      walSizeMb: 1,
-      ageSeconds: 1,
-    };
-
-    vi.mocked(stalenessMod.detectStalenessRisk).mockResolvedValue(risk);
+    vi.mocked(stalenessMod.detectStalenessRisk).mockResolvedValue(
+      STALENESS_RISK,
+    );
 
     const result = await findSimilar({ similarTo: SEED_KICK });
 
-    expect(result.stalenessRisk).toStrictEqual(risk);
+    expect(result.stalenessRisk).toStrictEqual(STALENESS_RISK);
   });
 
   it("reports a missing similarTo arg without throwing", async () => {
@@ -158,17 +153,8 @@ describe("findSimilar", () => {
   });
 
   it("degrades to dbAvailable:false when a query throws", async () => {
-    const broken = createBrokenLibraryDb();
-
-    try {
-      vi.mocked(dbPathMod.findLiveFilesDbPath).mockResolvedValue(broken.dbPath);
-
-      const result = await findSimilar({ similarTo: SEED_KICK });
-
-      expect(result.dbAvailable).toBe(false);
-      expect(result.reason).toContain("Failed to read Live database");
-    } finally {
-      broken.cleanup();
-    }
+    await expectQueryDegradesOnBrokenDb(dbPathMod, () =>
+      findSimilar({ similarTo: SEED_KICK }),
+    );
   });
 });
